0.0.2

+++ src/main/java/vavi/sound/dx7/FmOpKernel.java
@@ -27,7 +27,7 @@ class FmOpKernel {
             for (int i = 0; i < Note.N; i++) {
                 gain += dgain;
                 int y = Sin.lookup(phase + input[i]);
-                output[i] += ((long) y * (long) gain) >> 24;
+                output[i] += (int) (((long) y * (long) gain) >> 24);
                 phase += freq;
             }
         } else {
@@ -48,7 +48,7 @@ class FmOpKernel {
             for (int i = 0; i < Note.N; i++) {
                 gain += dgain;
                 int y = Sin.lookup(phase);
-                output[i] += ((long) y * (long) gain) >> 24;
+                output[i] += (int) (((long) y * (long) gain) >> 24);
                 phase += freq;
             }
         } else {
